# openapi/openapi.json
{
    "info": {
        "contact": {
            "email": "haleyhart123@gmail.com",
            "name": "Haley Hartsaw"
        },
        "description": "This app allows persons to schedule, edit or delete\nappointments made for vehicle maintenance at the Bodyshop\n",
        "title": "Bodyshop",
        "version": "1.0.0"
    },
    "openapi": "3.0.0",
    "paths": {
        "/appointments": {
            "get": {
                "description": "",
                "operationId": "appointment-list",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "appointments": [
                                        {
                                            "appointment_date": "",
                                            "appointment_id": "F889",
                                            "client": {
                                                "client_id": "11291"
                                            },
                                            "vehicle": {
                                                "vehicle_id": "2995"
                                            }
                                        },
                                        {
                                            "appointment_date": "",
                                            "appointment_id": "F142",
                                            "client": {
                                                "client_id": "11021"
                                            },
                                            "vehicle": {
                                                "vehicle_id": "6756"
                                            }
                                        }
                                    ]
                                },
                                "schema": {
                                    "description": "Appointment List",
                                    "properties": {
                                        "appointments": {
                                            "items": {
                                                "properties": {
                                                    "appointment_date": {
                                                        "type": "string"
                                                    },
                                                    "appointment_id": {
                                                        "type": "string"
                                                    },
                                                    "client": {
                                                        "items": {
                                                            "properties": {
                                                                "client_id": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "type": "array"
                                                    },
                                                    "vehicle": {
                                                        "items": {
                                                            "properties": {
                                                                "vehicle_id": {
                                                                    "type": "string"
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "type": "array"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "required": [
                                        "appointment_id"
                                    ],
                                    "title": "Appointment List",
                                    "type": "object"
                                }
                            }
                        },
                        "description": "List of all Appointments:"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "summary": "List of all Appointments.",
                "tags": [
                    "appointment"
                ]
            },
            "post": {
                "description": "",
                "operationId": "appointment-create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "appointment_id": {
                                        "type": "string"
                                    },
                                    "service_date": {
                                        "type": "string"
                                    },
                                    "service_id": {
                                        "type": "string"
                                    },
                                    "service_time": {
                                        "type": "string"
                                    },
                                    "vehicle_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    },
                    "description": "Create an Appointment"
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Appointment Scheduled",
                                    "properties": {
                                        "appointment_id": {
                                            "type": "string"
                                        }
                                    },
                                    "title": "Appointment Scheduled",
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Appointment Successfully Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "headers": {
                            "Content-Length": {
                                "description": "Appointment ID of appointment to be created.",
                                "required": true,
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Content-Type": {
                                "description": "Bad Type",
                                "required": true,
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Date": {
                                "description": "Appointment ID of appointment to be created.",
                                "required": true,
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "summary": "Schedule an appointment.",
                "tags": [
                    "appointment"
                ]
            }
        },
        "/appointments/{appointment_id}": {
            "delete": {
                "description": "",
                "operationId": "appointment-delete",
                "parameters": [
                    {
                        "description": "Appointment ID.",
                        "in": "path",
                        "name": "appointment_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Appointment Deleted",
                                    "properties": {
                                        "appointment_id": {
                                            "type": "string"
                                        }
                                    },
                                    "title": "Appointment Deleted",
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Appointment Successfully Deleted"
                    }
                },
                "summary": "Delete appointment.",
                "tags": [
                    "appointment"
                ]
            },
            "get": {
                "description": "",
                "operationId": "appointment",
                "parameters": [
                    {
                        "description": "Appointment ID.",
                        "in": "path",
                        "name": "appointment_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "appointment_date": "2019-01-15T09:00:00+0000",
                                    "appointment_id": "F889",
                                    "client": {
                                        "client_id": "11279"
                                    },
                                    "vehicle": {
                                        "vehicle_id": "2763"
                                    }
                                },
                                "schema": {
                                    "description": "Appointment Information",
                                    "properties": {
                                        "appointment_date": {
                                            "description": "Date and time of scheduled appointment",
                                            "format": "date-time",
                                            "type": "string"
                                        },
                                        "appointment_id": {
                                            "description": "Unique Appointment ID",
                                            "type": "string"
                                        },
                                        "client": {
                                            "properties": {
                                                "client_id": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "vehicle": {
                                            "properties": {
                                                "vehicle_id": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "required": [
                                        "appointment_id",
                                        "appointment_date",
                                        "client",
                                        "vehicle"
                                    ],
                                    "title": "Appointment Info",
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Appointment Information Retrieved"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "summary": "Get Appointment Information.",
                "tags": [
                    "appointment"
                ]
            }
        },
        "/appointments/{appointment_id}/services": {
            "get": {
                "description": "",
                "operationId": "appointment-scheduled-service",
                "parameters": [
                    {
                        "description": "Appointment ID.",
                        "in": "path",
                        "name": "appointment_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "services": [
                                        {
                                            "appointment_id": "F889",
                                            "service_id": "20",
                                            "service_type": "State Inspectin"
                                        }
                                    ]
                                },
                                "schema": {
                                    "description": "Appointment Services",
                                    "properties": {
                                        "services": {
                                            "items": {
                                                "properties": {
                                                    "appointment_id": {
                                                        "type": "string"
                                                    },
                                                    "service_id": {
                                                        "type": "string"
                                                    },
                                                    "service_type": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "title": "Appointment Services",
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Appointment Scheduled Services"
                    }
                },
                "summary": "Scheduled Services.",
                "tags": [
                    "appointment"
                ]
            }
        },
        "/clients": {
            "get": {
                "description": "",
                "operationId": "client-list",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "client": [
                                        {
                                            "client_id": "11291",
                                            "name": "Deku Midoriya",
                                            "phone": "(999)222-8374"
                                        },
                                        {
                                            "client_id": "11021",
                                            "name": "John Smith",
                                            "phone": "(555)676-9900"
                                        },
                                        {
                                            "client_id": "11010",
                                            "name": "Sally Jones",
                                            "phone": "(555)666-7777"
                                        }
                                    ]
                                },
                                "schema": {
                                    "description": "Client List",
                                    "properties": {
                                        "clients": {
                                            "items": {
                                                "properties": {
                                                    "client_id": {
                                                        "type": "string"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "phone": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "required": [
                                        "client_id"
                                    ],
                                    "title": "Client List",
                                    "type": "object"
                                }
                            }
                        },
                        "description": "List of all Clients:"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "summary": "List of all Clients.",
                "tags": [
                    "client"
                ]
            },
            "post": {
                "description": "",
                "operationId": "client-create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "type": "string"
                                    },
                                    "client_name": {
                                        "type": "string"
                                    },
                                    "client_phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    },
                    "description": "Create a Client"
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Client Created",
                                    "properties": {
                                        "client_id": {
                                            "type": "string"
                                        }
                                    },
                                    "title": "Client Created",
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Client Successfully Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "summary": "Create a client.",
                "tags": [
                    "client"
                ]
            }
        },
        "/clients/{client_id}": {
            "delete": {
                "description": "",
                "operationId": "client-delete",
                "parameters": [
                    {
                        "description": "Client ID.",
                        "in": "path",
                        "name": "client_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Client Deleted",
                                    "properties": {
                                        "client_id": {
                                            "type": "string"
                                        }
                                    },
                                    "title": "Client Deleted",
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Client Successfully Deleted"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "summary": "Delete a client.",
                "tags": [
                    "client"
                ]
            },
            "get": {
                "description": "",
                "operationId": "client",
                "parameters": [
                    {
                        "description": "Client ID.",
                        "in": "path",
                        "name": "client_id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "example": {
                                    "client": {
                                        "client_id": "11021",
                                        "name": "John Smith",
                                        "phone": "(555)676-9900"
                                    },
                                    "vehicle": [
                                        {
                                            "vehicle_id": "6124"
                                        },
                                        {
                                            "vehicle_id": "6756"
                                        }
                                    ]
                                },
                                "schema": {
                                    "description": "Client Information",
                                    "properties": {
                                        "client": {
                                            "properties": {
                                                "client_id": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "phone": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "vehicle": {
                                            "items": {
                                                "properties": {
                                                    "vehicle_id": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "required": [
                                        "client_id"
                                    ],
                                    "title": "Client Info",
                                    "type": "object"
                                }
                            }
                        },
                        "description": "Client Information Retrieved"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "summary": "Get Client Information.",
                "tags": [
                    "client"
                ]
            }
        },
...
    }
}